Update tests to use JUnit 5
The already used dependency "spring-boot-starter-test" provides JUnit 5.7.0. The dependency for JUnit 4.12 in the POM file is therefore not necessary or can even lead to conflicts because of the two different versions.

This change to use JUnit 5 in the code results in "mvn test" now actually recognizing the test classes. This was not the case before and could be due to the mixture of JUnit 4 and the JUnit 5 oriented packages in "spring-boot-starter-test".